# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "hiphouse API",
    "description": "크롤러가 채운 Postgres(artists/albums) CRUD 서버.",
    "version": "0.1.0"
  },
  "servers": [{ "url": "/" }],
...
  "paths": {
    "/albums": {
      "get": {
        "tags": ["albums"],
        "summary": "앨범 목록 (연도/아티스트 필터)",
        "parameters": [
          { "name": "year", "in": "query", "schema": { "type": "integer" }, "description": "발매 연도로 필터" },
          { "name": "artist_id", "in": "query", "schema": { "type": "string" }, "description": "아티스트 ID로 필터" },
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "검색어 — 앨범명(한글 표시 이름 포함), 크레딧 아티스트명(한글 표시 이름 포함), 아티스트 연관검색어(aliases), 트랙명에 매칭" },
          { "name": "deleted", "in": "query", "schema": { "type": "string", "enum": ["only"] }, "description": "only = 소프트삭제된 앨범만 (관리자 전용; 비관리자에겐 무시됨)" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 200 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": {
            "description": "앨범 배열",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Album" } } } }
          }
        }
      },
      "post": {
        "tags": ["albums"],
        "summary": "앨범 생성",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } }
        },
        "responses": {
          "201": { "description": "생성됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "description": "id 중복", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/albums/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "tags": ["albums"],
        "summary": "앨범 단건 조회",
        "responses": {
          "200": { "description": "앨범", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["albums"],
        "summary": "앨범 수정 (전체 교체)",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } }
        },
        "responses": {
          "200": { "description": "수정됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Album" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["albums"],
        "summary": "앨범 삭제",
        "responses": {
          "204": { "description": "삭제됨" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/albums/{id}/tracks": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "tags": ["albums"],
        "summary": "앨범 트랙 목록",
        "description": "disc_number, track_number 순 정렬. 아직 트랙 동기화 전인 앨범은 빈 배열.",
        "responses": {
          "200": { "description": "트랙 배열", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Track" } } } } }
        }
      }
    },
    "/albums/{id}/display-name": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "put": {
        "tags": ["albums"],
        "summary": "앨범 한글 표시 이름만 교체 (관리자 전용)",
        "description": "크롤러가 관리하는 필드는 건드리지 않는다. 빈 문자열/null이면 해제되어 Spotify 원본명으로 표시된다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "display_name": { "type": "string", "nullable": true, "example": "새삥" }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "수정됨" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/artists": {
      "get": {
        "tags": ["artists"],
        "summary": "아티스트 목록",
        "parameters": [
          { "name": "q", "in": "query", "schema": { "type": "string" }, "description": "검색어 — 아티스트명, 한글 표시 이름, 또는 연관검색어(aliases)에 매칭" },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50, "minimum": 1, "maximum": 200 } },
          { "name": "offset", "in": "query", "schema": { "type": "integer", "default": 0, "minimum": 0 } }
        ],
        "responses": {
          "200": {
            "description": "아티스트 배열",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Artist" } } } }
          }
        }
      },
      "post": {
        "tags": ["artists"],
        "summary": "아티스트 생성",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Artist" } } }
        },
        "responses": {
          "201": { "description": "생성됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Artist" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "409": { "description": "id 중복", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/artists/{id}": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "get": {
        "tags": ["artists"],
        "summary": "아티스트 단건 조회",
        "responses": {
          "200": { "description": "아티스트", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Artist" } } } },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["artists"],
        "summary": "아티스트 수정",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Artist" } } }
        },
        "responses": {
          "200": { "description": "수정됨", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Artist" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["artists"],
        "summary": "아티스트 삭제",
        "responses": {
          "204": { "description": "삭제됨" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/admin/spotify/artists": {
      "get": {
        "tags": ["artists"],
        "summary": "Spotify 아티스트 검색 (관리자 크롤링용)",
        "description": "Spotify 검색 후보에 우리 DB의 앨범 보유 수(albums_in_db)를 붙여 반환. key로 FIRST_/SECOND_ 자격증명 선택.",
        "parameters": [
          { "name": "q", "in": "query", "required": true, "schema": { "type": "string" } },
          { "name": "key", "in": "query", "schema": { "type": "string", "enum": ["first", "second"] } }
        ],
        "responses": {
          "200": {
            "description": "후보 배열",
            "content": { "application/json": { "schema": { "type": "array", "items": { "type": "object", "properties": {
              "id": { "type": "string" }, "name": { "type": "string" },
              "image_url": { "type": "string", "nullable": true },
              "followers": { "type": "integer" }, "albums_in_db": { "type": "integer" }
            } } } } }
          }
        }
      }
    },
    "/admin/spotify/crawl": {
      "post": {
        "tags": ["artists"],
        "summary": "아티스트 앨범 크롤 (관리자)",
        "description": "해당 아티스트의 전체 앨범을 Spotify에서 가져와 저장. 앨범 0개면 아티스트를 저장하지 않음. 크레딧 아티스트 중 이미지 없는 것만 개별 enrich. 트랙 미동기화 앨범(신규 포함)은 트랙까지 같이 수집.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "required": ["artist_id"], "properties": {
            "artist_id": { "type": "string" },
            "key": { "type": "string", "enum": ["first", "second"] }
          } } } }
        },
        "responses": {
          "200": { "description": "크롤 결과 {albums, saved, artist_name, artists, enriched, tracks_synced}" },
          "502": { "description": "Spotify API 오류", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/admin/tracks/backfill": {
      "post": {
        "tags": ["albums"],
        "summary": "트랙 백필 배치 실행 (관리자)",
        "description": "트랙 미동기화 앨범을 최신 발매순으로 최대 limit개 처리 (앨범당 Spotify 요청 1회, 앨범별 커밋). 404 앨범은 빈 트랙으로 완료 처리. 그 외 Spotify 오류는 배치를 중단하고 진행분과 함께 error로 보고. remaining이 0이 될 때까지 반복 호출.",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object", "properties": {
            "key": { "type": "string", "enum": ["first", "second"] },
            "limit": { "type": "integer", "default": 20, "minimum": 1, "maximum": 50 }
          } } } }
        },
        "responses": {
          "200": { "description": "{synced, not_found, tracks, remaining, albums:[{id,name,tracks,not_found}], error?}" }
        }
      }
    },
    "/artists/merge": {
      "post": {
        "tags": ["artists"],
        "summary": "아티스트 병합 (관리자 전용)",
        "description": "merged_ids의 앨범 크레딧을 master로 이동하고, 이름·연관검색어를 master의 연관검색어로 흡수한 뒤 중복 아티스트를 삭제한다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["master_id", "merged_ids"],
                "properties": {
                  "master_id": { "type": "string" },
                  "merged_ids": { "type": "array", "items": { "type": "string" } }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "병합된 master 아티스트", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Artist" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/artists/{id}/aliases": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "put": {
        "tags": ["artists"],
        "summary": "연관검색어(aliases)만 교체 (관리자 전용)",
        "description": "크롤러가 관리하는 필드(name/image/genres/spotify_url)는 건드리지 않고 aliases만 바꾼다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["aliases"],
                "properties": {
                  "aliases": { "type": "array", "items": { "type": "string" }, "example": ["블랙넛", "블넛"] }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "수정된 아티스트", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Artist" } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/artists/{id}/display-name": {
      "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
      "put": {
        "tags": ["artists"],
        "summary": "아티스트 한글 표시 이름만 교체 (관리자 전용)",
        "description": "크롤러가 관리하는 필드(name/image/genres/spotify_url)는 건드리지 않는다. 빈 문자열/null이면 해제되어 Spotify 원본명으로 표시된다.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "display_name": { "type": "string", "nullable": true, "example": "지코" }
                }
              }
            }
          }
        },
        "responses": {
          "204": { "description": "수정됨" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    }
  },
  "components": {
    "responses": {
      "BadRequest": { "description": "잘못된 요청", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "없음", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Album": {
        "type": "object",
        "required": ["id", "name", "artists"],
        "properties": {
          "id": { "type": "string", "example": "4aawyAB9vmqN3uQ7FjRGTy" },
          "name": { "type": "string", "example": "Album Name" },
          "display_name": { "type": "string", "nullable": true, "example": "새삥", "description": "관리자가 등록한 한글 표시 이름. 없으면 name으로 표시하고, 검색 시 name과 함께 매칭됨" },
          "artists": {
            "type": "array",
            "description": "크레딧된 모든 아티스트 (Spotify 순서). album_artists 조인에서 집계.",
            "items": { "$ref": "#/components/schemas/Artist" }
          },
          "release_date": { "type": "string", "nullable": true, "example": "2024-05-01" },
          "year": { "type": "integer", "nullable": true, "example": 2024 },
          "album_type": { "type": "string", "nullable": true, "example": "album" },
          "total_tracks": { "type": "integer", "nullable": true, "example": 12 },
          "image_url": { "type": "string", "nullable": true },
          "spotify_url": { "type": "string", "nullable": true }
        }
      },
      "Artist": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": { "type": "string", "example": "1uNFoZAHBGtllmzznpCI3s" },
          "name": { "type": "string", "nullable": true, "example": "Artist Name" },
          "display_name": { "type": "string", "nullable": true, "example": "지코", "description": "관리자가 등록한 한글 표시 이름. 없으면 name으로 표시하고, 검색 시 name과 함께 매칭됨" },
          "image_url": { "type": "string", "nullable": true, "description": "Spotify 아티스트 사진 (없을 수 있음)" },
          "genres": { "type": "array", "items": { "type": "string" }, "nullable": true, "description": "Spotify 아티스트 장르 (비어있을 수 있음)" },
          "spotify_url": { "type": "string", "nullable": true, "description": "Spotify 아티스트 프로필 URL" },
          "aliases": { "type": "array", "items": { "type": "string" }, "nullable": true, "description": "연관검색어 — 관리자가 등록한 검색 키워드 (예: 영문명의 한글 표기). 검색 시 name과 함께 매칭됨" },
          "followers": { "type": "integer", "nullable": true, "description": "Spotify 팔로워 수 — 크롤러/enrich가 채움 (아직 수집 전이면 null). 화면 표시용이 아니라 관리자 enrich 판단용" },
          "follower_count": { "type": "integer", "description": "서비스 내부 팔로워 수 — follows 테이블 집계" }
        }
      },
      "Track": {
        "type": "object",
        "required": ["id", "disc_number", "track_number", "name", "explicit", "artists"],
        "properties": {
          "id": { "type": "string", "example": "3n3Ppam7vgaVa1iaRUc9Lp" },
          "disc_number": { "type": "integer", "example": 1 },
          "track_number": { "type": "integer", "example": 3 },
          "name": { "type": "string", "example": "Track Name" },
          "duration_ms": { "type": "integer", "nullable": true, "example": 207959 },
          "explicit": { "type": "boolean" },
          "spotify_url": { "type": "string", "nullable": true },
          "artists": {
            "type": "array",
            "description": "참여 아티스트 (피처링 포함, Spotify 크레딧 순서)",
            "items": { "type": "object", "properties": { "id": { "type": "string" }, "name": { "type": "string" } } }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": { "error": { "type": "string" } }
      }
    }
  }
}